+++ src/main/java/course/concurrency/m3_shared/immutable/Order.java
@@ -2,36 +2,56 @@ package course.concurrency.m3_shared.immutable;
 
 import java.util.List;
 
-import static course.concurrency.m3_shared.immutable.Order.Status.NEW;
+import static course.concurrency.m3_shared.immutable.Order.Status.*;
 
 public class Order {
 
     public enum Status { NEW, IN_PROGRESS, DELIVERED }
 
-    private Long id;
-    private List<Item> items;
-    private PaymentInfo paymentInfo;
-    private boolean isPacked;
-    private Status status;
+    private final Long id;
+    private final List<Item> items;
+    private final PaymentInfo paymentInfo;
+    private final boolean isPacked;
+    private final Status status;
 
-    public Order(List<Item> items) {
-        this.items = items;
+    public Order(Long id, List<Item> items) {
+        this.id = id;
+        this.items = List.copyOf(items);
+        this.paymentInfo = null;
+        this.isPacked = false;
         this.status = NEW;
     }
 
-    public synchronized boolean checkStatus() {
-        if (items != null && !items.isEmpty() && paymentInfo != null && isPacked) {
-            return true;
-        }
-        return false;
+    public Order(Long id, List<Item> items, PaymentInfo paymentInfo, boolean isPacked, Status status) {
+        this.id = id;
+        this.items = List.copyOf(items);
+        this.paymentInfo = paymentInfo;
+        this.isPacked = isPacked;
+        this.status = status;
+    }
+
+    public static Order packed(Order order) {
+        return new Order(order.getId(), order.getItems(), order.getPaymentInfo(),
+                true, IN_PROGRESS);
+    }
+
+    public static Order payed(Order order, PaymentInfo paymentInfo) {
+        return new Order(order.getId(), order.getItems(), paymentInfo,
+                order.isPacked, IN_PROGRESS);
+    }
+
+    public static Order delivered(Order order) {
+        return new Order(order.getId(), order.getItems(), order.getPaymentInfo(),
+                order.isPacked(), DELIVERED);
+    }
+
+    public boolean isReadyForDelivery() {
+        return items != null && !items.isEmpty() && paymentInfo != null && isPacked
+                && !status.equals(DELIVERED);
     }
 
     public Long getId() {
         return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
     }
 
     public List<Item> getItems() {
@@ -42,25 +62,12 @@ public class Order {
         return paymentInfo;
     }
 
-    public void setPaymentInfo(PaymentInfo paymentInfo) {
-        this.paymentInfo = paymentInfo;
-        this.status = Status.IN_PROGRESS;
-    }
-
     public boolean isPacked() {
         return isPacked;
-    }
-
-    public void setPacked(boolean packed) {
-        isPacked = packed;
-        this.status = Status.IN_PROGRESS;
     }
 
     public Status getStatus() {
         return status;
     }
 
-    public void setStatus(Status status) {
-        this.status = status;
-    }
 }