+++ src/test/java/course/concurrency/m3_shared/immutable/OrderServiceTests.java
@@ -70,18 +70,18 @@ public class OrderServiceTests {
         int iterations = 10_000;
         ExecutorService executor = Executors.newFixedThreadPool(Runtime.getRuntime().availableProcessors());
 
-        BlockingQueue<Long> orderToPack = new ArrayBlockingQueue<>(iterations*3);
+        BlockingQueue<Long> orderToPack = new ArrayBlockingQueue<>(iterations * 3);
         BlockingQueue<Long> orderToPay = new ArrayBlockingQueue<>(iterations);
         BlockingQueue<Long> orderIdLog = new ArrayBlockingQueue<>(iterations);
 
-        for(int i = 0; i< iterations; ++i) {
+        for (int i = 0; i < iterations; ++i) {
             Long id = service.createOrder(items);
-                        orderToPack.offer(id);
-                        orderToPay.offer(id);
-                        orderIdLog.offer(id);
+            orderToPack.offer(id);
+            orderToPay.offer(id);
+            orderIdLog.offer(id);
         }
 
-        for(int i = 0; i< iterations; ++i) {
+        for (int i = 0; i < iterations; ++i) {
             executor.submit(() -> service.setPacked(orderToPack.poll()));
             executor.submit(() -> service.updatePaymentInfo(orderToPay.poll(), new PaymentInfo()));
         }